// src/interfaces/orderbook.ts
export interface Orders {
  symbol: string;
  ts: number;
  asks: number[][];
  bids: number[][];
}

export interface MarketTrade {
  symbol: string;
  price: number;
  size: number;
  side: "BUY" | "SELL";
  ts: number;
}

interface Ask {
  price: number;
  quantity: number;
}
export interface RequestOrder {
  asks: Ask[];
  bids: Ask[];
  timestamp: number;
}

export interface Trade {
  symbol: string;
  side: "BUY" | "SELL";
  executed_price: number;
  executed_quantity: number;
  executed_timestamp: number;
}

export interface TokenMetadata {
  id: string;
  name: string;
  symbol: string;
  decimals: number;
  icon: string;
}

export interface Ticker {
  symbol: string;
  open: number;
  close: number;
  high: number;
  low: number;
  volume: number;
  amount: number;
  count: number;
  symbol_from_token_meta?: TokenMetadata;
  symbol_to_token_meta?: TokenMetadata;
}
// ...
export interface MarkPrice {
  symbol: string;
  price: number;
}

export interface IndexPrice {
  symbol: string;
  price: number;
}

export interface EstFundingrate {
  fundingRate: number;
  fundingTs: number;
  symbol: string;
}
// ...
export interface OpenInterest {
  symbol: string;
  openInterest: number;
}